feat: separate TS into two files to make json-schema generator easier
