+++ examples/rust-status-message/index.schema.json
@@ -1,0 +1,76 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "definitions": {
+    "AccountId": {
+      "maxLength": 64,
+      "minLength": 2,
+      "pattern": "^(([a-z\\d]+[-_])*[a-z\\d]+\\.)*([a-z\\d]+[-_])*[a-z\\d]+$",
+      "type": "string"
+    },
+    "Balance": {
+      "$ref": "#/definitions/U128",
+      "description": "Balance is a type for storing amounts of tokens, specified in yoctoNEAR."
+    },
+    "Base64VecU8": {
+      "description": "base64 string.",
+      "type": "string"
+    },
+    "Duration": {
+      "$ref": "#/definitions/u64",
+      "description": "Raw type for duration in nanoseconds"
+    },
+    "Gas": {
+      "$ref": "#/definitions/u64",
+      "description": "Represents the amount of NEAR tokens in \"gas units\" which are used to fund transactions."
+    },
+    "GetStatus": {
+      "additionalProperties": false,
+      "contractMethod": "view",
+      "description": "Retreive a message for a given account id",
+      "properties": {
+        "account_id": {
+          "$ref": "#/definitions/AccountId"
+        }
+      },
+      "required": [
+        "account_id"
+      ],
+      "type": "object"
+    },
+    "PublicKey": {
+      "description": "Public key in a binary format with base58 string serialization with human-readable curve. The key types currently supported are `secp256k1` and `ed25519`.\n\nEd25519 public keys accepted are 32 bytes and secp256k1 keys are the uncompressed 64 format.",
+      "type": "string"
+    },
+    "SetStatus": {
+      "additionalProperties": false,
+      "contractMethod": "change",
+      "description": "Store a message for current signer account",
+      "properties": {
+        "message": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "message"
+      ],
+      "type": "object"
+    },
+    "StorageUsage": {
+      "$ref": "#/definitions/u64"
+    },
+    "Timestamp": {
+      "$ref": "#/definitions/u64",
+      "description": "Raw type for timestamp in nanoseconds"
+    },
+    "U128": {
+      "description": "String representation of a u128-bit integer",
+      "pattern": "^[0-9]+$\nNote: largest u128 is \"340282366920938463463374607431768211455\"",
+      "type": "string"
+    },
+    "u64": {
+      "maximum": 18446744073709552000,
+      "minimum": 0,
+      "type": "integer"
+    }
+  }
+}
