+++ examples/rust-status-message/ts/index.ts
@@ -1,88 +1,22 @@
-import { Account, transactions, providers, DEFAULT_FUNCTION_CALL_GAS } from 'near-api-js';
-
-
-import BN from 'bn.js';
-export interface ChangeMethodOptions {
-  gas?: BN;
-  attachedDeposit?: BN;
-  walletMeta?: string;
-  walletCallbackUrl?: string;
-}
-export interface ViewFunctionOptions {
-  // TODO currently JSON schema generator doesn't like function types
-  parse?: any;
-  // TODO currently JSON schema generator doesn't like function types
-  stringify?: any;
-}
-
-/** 
-* @minimum 0
-* @maximum 18446744073709551615
-* @asType integer
-*/
-export type u64 = number;
-/** 
-* @minimum -9223372036854775808
-* @maximum 9223372036854775807
-* @asType integer
-*/
-export type i64 = number;
-
-/**
-* @minimum  0 
-* @maximum 255
-* @asType integer
-* */
-export type u8 = number;
-/**
-* @minimum  -128 
-* @maximum 127
-* @asType integer
-* */
-export type i8 = number;
-/**
-* @minimum  0 
-* @maximum 65535
-* @asType integer
-* */
-export type u16 = number;
-/**
-* @minimum -32768 
-* @maximum 32767
-* @asType integer
-* */
-export type i16 = number;
-/**
-* @minimum 0 
-* @maximum 4294967295
-* @asType integer
-* */
-export type u32 = number;
-/**
-* @minimum 0 
-* @maximum 4294967295
-* @asType integer
-* */
-export type usize = number;
-/**
-* @minimum  -2147483648 
-* @maximum 2147483647
-* @asType integer
-* */
-export type i32 = number;
-
-/**
-* @minimum -3.40282347E+38
-* @maximum 3.40282347E+38
-*/
-export type f32 = number;
-
-/**
-* @minimum -1.7976931348623157E+308
-* @maximum 1.7976931348623157E+308
-*/
-export type f64 = number;
-/**
+import {
+  Account,
+  transactions,
+  providers,
+  DEFAULT_FUNCTION_CALL_GAS,
+  u8,
+  i8,
+  u16,
+  i16,
+  u32,
+  i32,
+  u64,
+  i64,
+  f32,
+  f64,
+  BN,
+  ChangeMethodOptions,
+  ViewFunctionOptions,
+} from './helper';/**
 * StorageUsage is used to count the amount of storage used by a contract.
 */
 export type StorageUsage = u64;
@@ -130,38 +64,52 @@ export class Contract {
   
   constructor(public account: Account, public readonly contractId: string){}
   
-  async set_status(args: {
-    message: string;
-  }, options?: ChangeMethodOptions): Promise<void> {
-    return providers.getTransactionLastResult(await this.set_statusRaw(args, options));
-  }
-  set_statusRaw(args: {
-    message: string;
-  }, options?: ChangeMethodOptions): Promise<providers.FinalExecutionOutcome> {
-    return this.account.functionCall({contractId: this.contractId, methodName: "set_status", args, ...options});
-  }
-  set_statusTx(args: {
-    message: string;
-  }, options?: ChangeMethodOptions): transactions.Action {
-    return transactions.functionCall("set_status", args, options?.gas ?? DEFAULT_FUNCTION_CALL_GAS, options?.attachedDeposit ?? new BN(0))
-  }
+  /**
+  * Retreive a message for a given account id
+  */
   get_status(args: {
     account_id: AccountId;
   }, options?: ViewFunctionOptions): Promise<string | null> {
     return this.account.viewFunction(this.contractId, "get_status", args, options);
   }
+  /**
+  * Store a message for current signer account
+  */
+  async set_status(args: {
+    message: string;
+  }, options?: ChangeMethodOptions): Promise<void> {
+    return providers.getTransactionLastResult(await this.set_statusRaw(args, options));
+  }
+  /**
+  * Store a message for current signer account
+  */
+  set_statusRaw(args: {
+    message: string;
+  }, options?: ChangeMethodOptions): Promise<providers.FinalExecutionOutcome> {
+    return this.account.functionCall({contractId: this.contractId, methodName: "set_status", args, ...options});
+  }
+  /**
+  * Store a message for current signer account
+  */
+  set_statusTx(args: {
+    message: string;
+  }, options?: ChangeMethodOptions): transactions.Action {
+    return transactions.functionCall("set_status", args, options?.gas ?? DEFAULT_FUNCTION_CALL_GAS, options?.attachedDeposit ?? new BN(0))
+  }
 }
 /**
-* 
-* @contractMethod change
-*/
-export interface SetStatus {
-  message: string;
-}
-/**
+* Retreive a message for a given account id
 * 
 * @contractMethod view
 */
 export interface GetStatus {
   account_id: AccountId;
+}
+/**
+* Store a message for current signer account
+* 
+* @contractMethod change
+*/
+export interface SetStatus {
+  message: string;
 }
