+++ examples/rust-status-message/tsconfig.json
@@ -66,5 +66,5 @@
     "skipLibCheck": true,                     /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */
   },
-  "include": ["index.ts"]
+  "include": ["ts/*"]
 }
