+++ package.json
@@ -1,5 +1,6 @@
 {
   "devDependencies": {
+    "@types/bn.js": "^5.1.0",
     "@types/node": "^15.12.2",
     "@typescript-eslint/eslint-plugin": "^4.27.0",
     "@typescript-eslint/parser": "^4.27.0",
